meta-data work moved to ungridded_data_metadata; deprecation of several unused methods
