+++ pyaerocom/ungridded_data_metadata.py
@@ -1,12 +1,19 @@
-from copy import deepcopy
+import fnmatch
 import logging
 import sys
+from copy import deepcopy
+from datetime import datetime
 from typing import Any
 
 import numpy as np
-from pyaerocom.exceptions import MetaDataError
-from pyaerocom.ungridded_data import UngriddedDataContainer
 
+from pyaerocom import const
+from pyaerocom.exceptions import DataCoverageError, MetaDataError, StationNotFoundError
+from pyaerocom.helpers import isnumeric, start_stop_str
+from pyaerocom.helpers_landsea_masks import get_mask_value, load_region_mask_xr
+from pyaerocom.mathutils import in_range
+from pyaerocom.ungridded_data import UngriddedDataContainer
+from pyaerocom.units_helpers import get_unit_conversion_fac
 
 logger = logging.getLogger(__name__)
 
@@ -44,6 +51,7 @@ class UngriddedDataMetadata(UngriddedDataContainer):
         other.filter_hist = deepcopy(self.filter_hist)
         other.is_vertical_profile = self._is_vertical_profile
 
+    @override
     def _get_data_revision_helper(self, data_id):
         """
         Helper method to get last data revision
@@ -210,7 +218,7 @@ class UngriddedDataMetadata(UngriddedDataContainer):
     @override
     def unique_station_names(self):
         """List of unique station names"""
-        return sorted(list(set(self.station_name)))
+        return sorted(set(self.station_name))
 
     @property
     @override
@@ -249,3 +257,583 @@ class UngriddedDataMetadata(UngriddedDataContainer):
         if not self.is_filtered:
             raise AttributeError("No filters were applied so far")
         return self.filter_hist[max(self.filter_hist)]
+
+    @override
+    def find_station_meta_indices(self, station_name_or_pattern, allow_wildcards=True):
+        """Find indices of all metadata blocks matching input station name
+
+        You may also use wildcard pattern as input (e.g. *Potenza*)
+
+        Parameters
+        ----------
+        station_pattern : str
+            station name or wildcard pattern
+        allow_wildcards : bool
+            if True, input station_pattern will be used as wildcard pattern and
+            all matches are returned.
+
+        Returns
+        -------
+        list
+           list containing all metadata indices that match the input station
+           name or pattern
+
+        Raises
+        ------
+        StationNotFoundError
+            if no such station exists in this data object
+        """
+        if not allow_wildcards:
+
+            def compare(x, y):
+                return fnmatch.fnmatch(x, y)
+
+        else:
+
+            def compare(x, y):
+                return x == y
+
+        idx = []
+        for i, meta in self.metadata.items():
+            if compare(meta["station_name"], station_name_or_pattern):
+                idx.append(i)
+        if len(idx) == 0:
+            raise StationNotFoundError(
+                f"No station available in UngriddedData that matches name {station_name_or_pattern}"
+            )
+        return idx
+
+    @override
+    def check_unit(self, var_name, unit=None):
+        """Check if variable unit corresponds to AeroCom unit
+
+        Parameters
+        ----------
+        var_name : str
+            variable name for which unit is to be checked
+        unit : :obj:`str`, optional
+            unit to be checked, if None, AeroCom default unit is used
+
+        Raises
+        ------
+        MetaDataError
+            if unit information is not accessible for input variable name
+        """
+        if unit is None:
+            unit = const.VARS[var_name]["units"]
+
+        units = []
+        for i, meta in self.metadata.items():
+            if var_name in meta["var_info"]:
+                try:
+                    u = meta["var_info"][var_name]["units"]
+                    if u not in units:
+                        units.append(u)
+                except KeyError:
+                    add_str = ""
+                    if "unit" in meta["var_info"][var_name]:
+                        add_str = (
+                            "Corresponding var_info dict contains "
+                            'attr. "unit", which is deprecated, please '
+                            "check corresponding reading routine. "
+                        )
+                    raise MetaDataError(
+                        f"Failed to access unit information for variable {var_name} "
+                        f"in metadata block {i}. {add_str}"
+                    )
+        if len(units) == 0 and str(unit) != "1":
+            raise MetaDataError(
+                f"Failed to access unit information for variable {var_name}. "
+                f"Expected unit {unit}"
+            )
+        for u in units:
+            if not get_unit_conversion_fac(u, unit, var_name) == 1:
+                raise MetaDataError(f"Invalid unit {u} detected (expected {unit})")
+
+    def _add_to_filter_history(self, info):
+        """Add info to :attr:`filter_hist`
+
+        Key is current system time string
+
+        Parameter
+        ---------
+        info
+            information to be appended to filter history
+        """
+        time_str = datetime.now().strftime("%Y%m%d%H%M%S")
+        self.filter_hist[int(time_str)] = info
+
+    @property
+    @override
+    def station_coordinates(self):
+        """dictionary with station coordinates
+
+        Returns
+        -------
+        dict
+            dictionary containing station coordinates (latitude, longitude,
+            altitude -> values) for all stations (keys) where these parameters
+            are accessible.
+        """
+        d = {"station_name": [], "latitude": [], "longitude": [], "altitude": []}
+
+        for i, meta in self.metadata.items():
+            if "station_name" not in meta:
+                logger.debug(f"Skipping meta-block {i}: station_name is not defined")
+                continue
+            elif not all(name in meta for name in const.STANDARD_COORD_NAMES):
+                logger.debug(
+                    f"Skipping meta-block {i} (station {meta['station_name']}): "
+                    f"one or more of the coordinates is not defined"
+                )
+                continue
+
+            stat = meta["station_name"]
+
+            if stat in d["station_name"]:
+                continue
+            d["station_name"].append(stat)
+            for k in const.STANDARD_COORD_NAMES:
+                d[k].append(meta[k])
+        return d
+
+    @property
+    def countries_available(self):
+        """
+        Alphabetically sorted list of country names available
+        """
+        # self.check_set_country()
+        countries = []
+        for idx, meta in self.metadata.items():
+            try:
+                countries.append(meta["country"])
+            except Exception:
+                logger.warning("No country information in meta block", idx)
+        if len(countries) == 0:
+            logger.warning(
+                "None of the metadata blocks contains "
+                "country information. You may want to "
+                "run class method check_set_country first "
+                "to automatically assign countries."
+            )
+        return sorted(set(countries))
+
+    def _init_meta_filters(self, **filter_attributes):
+        """Init filter dictionary for :func:`apply_filter_meta`
+
+        Parameters
+        ----------
+        **filter_attributes
+            valid meta keywords that are supposed to be filtered and the
+            corresponding filter values (or value ranges)
+            Only valid meta keywords are considered (e.g. data_id,
+            longitude, latitude, altitude, ts_type)
+
+        Returns
+        -------
+        tuple
+            3-element tuple containing
+
+            - dict: string match filters for metakeys \
+              (e.g. dict['data_id'] = 'AeronetSunV2Lev2.daily')
+            - dict: in-list match filters for metakeys \
+              (e.g. dict['station_name'] = ['stat1', 'stat2', 'stat3'])
+            - dict: in-range dictionary for metakeys \
+              (e.g. dict['longitude'] = [-30, 30])
+
+        """
+        # initiate filters that are checked
+        valid_keys = list(self.metadata[self.first_meta_idx])
+        str_f = {}
+        list_f = {}
+        range_f = {}
+        val_f = {}
+        for key, val in filter_attributes.items():
+            if key not in valid_keys:
+                raise OSError(
+                    f"Invalid input parameter for filtering: {key}. "
+                    f"Please choose from {valid_keys}"
+                )
+
+            if isinstance(val, str):
+                str_f[key] = val
+            elif isnumeric(val):
+                val_f[key] = val
+            elif isinstance(val, list | np.ndarray | tuple):
+                if all([isinstance(x, str) for x in val]):
+                    list_f[key] = val
+                elif len(val) == 2 and all([isnumeric(x) for x in val]):
+                    try:
+                        low, high = float(val[0]), float(val[1])
+                        if not low < high:
+                            raise ValueError("First entry needs to be smaller than 2nd")
+                        range_f[key] = [low, high]
+                    except Exception:
+                        list_f[key] = val
+                else:
+                    list_f[key] = val
+        return (str_f, list_f, range_f, val_f)
+
+    def _find_meta_matches(self, negate=None, *filters):
+        """Find meta matches for input attributes
+
+        Parameters
+        ----------
+        negate : list or str, optional
+            specified meta key(s) provided in `*filters` that are
+            supposed to be treated as 'not valid'. E.g. if
+            `station_name="bad_site"` is input in `filter_attributes` and if
+            `station_name` is listed in `negate`, then all metadata blocks
+            containing "bad_site" as station_name will be excluded in output
+            data object.
+        *filters
+            list of filters to be applied
+
+        Returns
+        -------
+        tuple
+            list of metadata indices that match input filter
+        """
+        if negate is None:
+            negate = []
+        elif isinstance(negate, str):
+            negate = [negate]
+        elif not isinstance(negate, list):
+            raise ValueError(f"Invalid input for negate {negate}, need list or str or None")
+        meta_matches = []
+        totnum = 0
+        for meta_idx, meta in self.metadata.items():
+            if self._check_filter_match(meta, negate, *filters):
+                meta_matches.append(meta_idx)
+                for var in meta["var_info"]:
+                    if var in self.ALLOWED_VERT_COORD_TYPES:
+                        continue  # altitude is not actually a variable but is stored in var_info like one
+                    try:
+                        totnum += len(self.meta_idx[meta_idx][var])
+                    except KeyError:
+                        logger.debug(
+                            f"Ignoring variable {var} in meta block {meta_idx} "
+                            f"since no data could be found"
+                        )
+
+        return (meta_matches, totnum)
+
+    @override
+    def filter_by_meta(self, negate=None, **filter_attributes):
+        """Flexible method to filter these data based on input meta specs
+
+        Parameters
+        ----------
+        negate : list or str, optional
+            specified meta key(s) provided via `filter_attributes` that are
+            supposed to be treated as 'not valid'. E.g. if
+            `station_name="bad_site"` is input in `filter_attributes` and if
+            `station_name` is listed in `negate`, then all metadata blocks
+            containing "bad_site" as station_name will be excluded in output
+            data object.
+        **filter_attributes
+            valid meta keywords that are supposed to be filtered and the
+            corresponding filter values (or value ranges)
+            Only valid meta keywords are considered (e.g. data_id,
+            longitude, latitude, altitude, ts_type)
+
+        Returns
+        -------
+        UngriddedData
+            filtered ungridded data object
+
+        Raises
+        ------
+        NotImplementedError
+            if attempt variables are supposed to be filtered (not yet possible)
+        IOError
+            if any of the input keys are not valid meta key
+
+        Example
+        -------
+        >>> import pyaerocom as pya
+        >>> r = pya.io.ReadUngridded(['AeronetSunV3Lev2.daily'], 'od550aer')
+        >>> data = r.read()
+        >>> data_filtered = data.filter_by_meta(data_id='AeronetSunV3Lev2.daily',
+        ...                                     longitude=[-30, 30],
+        ...                                     latitude=[20, 70],
+        ...                                     altitude=[0, 1000])
+        """
+
+        if "variables" in filter_attributes:
+            raise NotImplementedError("Cannot yet filter by variables")
+
+        # separate filters by strin, list, etc.
+        filters = self._init_meta_filters(**filter_attributes)
+
+        # find all metadata blocks that match the filters
+        meta_matches, totnum_new = self._find_meta_matches(
+            negate,
+            *filters,
+        )
+        if len(meta_matches) == len(self.metadata):
+            logger.info(f"Input filters {filter_attributes} result in unchanged data object")
+            return self
+        new = self._new_from_meta_blocks(meta_matches, totnum_new)
+        time_str = datetime.now().strftime("%Y%m%d%H%M%S")
+        new.filter_hist[int(time_str)] = filter_attributes
+        return new
+
+    @override
+    def filter_by_projection(
+        self, projection, xrange: tuple[float, float], yrange: tuple[float, float]
+    ):
+        """Filter the ungridded data to a horizontal bounding box given by a projection
+
+        :param projection: a function turning projection(lat, lon) -> (x, y)
+        :param xrange: x range (min/max included) in the projection plane
+        :param yrange: y range (min/max included) in the projection plane
+        """
+        meta_matches = []
+        totnum = 0
+        for meta_idx, meta in self.metadata.items():
+            lon = meta["longitude"]
+            lat = meta["latitude"]
+            x, y = projection(lat, lon)
+
+            match_x = in_range(x, xrange[0], xrange[1])
+            match_y = in_range(y, yrange[0], yrange[1])
+
+            if match_x and match_y:
+                meta_matches.append(meta_idx)
+                for var in meta["var_info"]:
+                    if var in self.ALLOWED_VERT_COORD_TYPES:
+                        continue  # altitude is not actually a variable but is stored in var_info like one
+                    try:
+                        totnum += len(self.meta_idx[meta_idx][var])
+                    except KeyError:
+                        logger.debug(
+                            f"Ignoring variable {var} in meta block {meta_idx} "
+                            f"since no data could be found"
+                        )
+
+        if len(meta_matches) == len(self.metadata):
+            logger.info("filter_by_projection result in unchanged data object")
+            return self
+        new = self._new_from_meta_blocks(meta_matches, totnum)
+        return new
+
+    @override
+    def apply_region_mask(self, region_id=None):
+        """
+        TODO : Write documentations
+
+        Parameters
+        ----------
+        region_id : str or list (of strings)
+            ID of region or IDs of multiple regions to be combined
+        """
+        if region_id not in const.HTAP_REGIONS:
+            raise ValueError(
+                f"Invalid input for region_id: {region_id}, choose from: {const.HTAP_REGIONS}"
+            )
+
+        # 1. find matches -> list of meta indices that are in region
+        # 2. Get total number of datapoints -> defines shape of output UngriddedData
+        # 3. Create
+
+        mask = load_region_mask_xr(region_id)
+
+        meta_matches = []
+        totnum = 0
+        for meta_idx, meta in self.metadata.items():
+            lon, lat = meta["longitude"], meta["latitude"]
+
+            mask_val = get_mask_value(lat, lon, mask)
+            if mask_val >= 1:  # coordinate is in mask
+                meta_matches.append(meta_idx)
+                for var in meta["var_info"]:
+                    totnum += len(self.meta_idx[meta_idx][var])
+
+        new = self._new_from_meta_blocks(meta_matches, totnum)
+        time_str = datetime.now().strftime("%Y%m%d%H%M%S")
+        new.filter_hist[int(time_str)] = f"Applied mask {region_id}"
+        new._check_index()
+        return new
+
+    def _meta_to_lists(self):
+        """List metadata for plotting
+
+        :return: list of metadata values
+        """
+        meta = {k: [] for k in self.metadata[self.first_meta_idx]}
+        for meta_item in self.metadata.values():
+            for k, v in meta.items():
+                v.append(meta_item[k])
+        return meta
+
+    def plot_station_timeseries(
+        self,
+        station_name,
+        var_name,
+        start=None,
+        stop=None,
+        ts_type=None,
+        insert_nans=True,
+        ax=None,
+        **kwargs,
+    ):  # pragma: no cover
+        """Plot time series of station and variable
+
+        Parameters
+        ----------
+        station_name : :obj:`str` or :obj:`int`
+            station name or index of station in metadata dict
+        var_name : str
+            name of variable to be retrieved
+        start
+            start time (optional)
+        stop
+            stop time (optional). If start time is provided and stop time not,
+            then only the corresponding year inferred from start time will be
+            considered
+        ts_type : :obj:`str`, optional
+            temporal resolution
+
+        **kwargs
+            Addifional keyword args passed to method :func:`pandas.Series.plot`
+
+        Returns
+        -------
+        axes
+            matplotlib axes instance
+
+        """
+        if ax is None:
+            import matplotlib.pyplot as plt
+            from pyaerocom.plot.config import FIGSIZE_DEFAULT
+
+            fig, ax = plt.subplots(figsize=FIGSIZE_DEFAULT)
+
+        stat = self.to_station_data(
+            station_name,
+            var_name,
+            start,
+            stop,
+            freq=ts_type,
+            merge_if_multi=True,
+            insert_nans=insert_nans,
+        )
+        ax = stat.plot_timeseries(var_name, ax=ax, **kwargs)
+        return ax
+
+    def plot_station_coordinates(
+        self,
+        var_name=None,
+        start=None,
+        stop=None,
+        ts_type=None,
+        color="r",
+        marker="o",
+        markersize=8,
+        fontsize_base=10,
+        legend=True,
+        add_title=True,
+        **kwargs,
+    ):  # pragma: no cover
+        """Plot station coordinates on a map
+
+        All input parameters are optional and may be used to add constraints
+        related to which stations are plotted. Default is all stations of all
+        times.
+
+        Parameters
+        ----------
+
+        var_name : :obj:`str`, optional
+            name of variable to be retrieved
+        start
+            start time (optional)
+        stop
+            stop time (optional). If start time is provided and stop time not,
+            then only the corresponding year inferred from start time will be
+            considered
+        ts_type : :obj:`str`, optional
+            temporal resolution
+        color : str
+            color of stations on map
+        marker : str
+            marker type of stations
+        markersize : int
+            size of station markers
+        fontsize_base : int
+            basic fontsize
+        legend : bool
+            if True, legend is added
+        add_title : bool
+            if True, title will be added
+        **kwargs
+            Addifional keyword args passed to
+            :func:`pyaerocom.plot.plot_coordinates`
+
+        Returns
+        -------
+        axes
+            matplotlib axes instance
+
+        """
+        from pyaerocom.plot.plotcoordinates import plot_coordinates
+
+        if len(self.contains_datasets) > 1:
+            logger.warning(
+                "UngriddedData object contains more than one "
+                "dataset ({}). Station coordinates will not be "
+                "distinguishable. You may want to apply a filter "
+                "first and plot them separately"
+            )
+
+        subset = self
+        if var_name is None:
+            info_str = "AllVars"
+        else:
+            if not isinstance(var_name, str):
+                raise ValueError("Can only handle single variable (or all -> input var_name=None)")
+            elif var_name not in subset.contains_vars:
+                raise ValueError(f"Input variable {var_name} is not available in dataset ")
+            info_str = var_name
+
+        try:
+            info_str += f"_{start_stop_str(start, stop, ts_type)}"
+        except Exception:
+            info_str += "_AllTimes"
+        if ts_type is not None:
+            info_str += f"_{ts_type}"
+
+        if all([x is None for x in (var_name, start, stop)]):  # use all stations
+            all_meta = subset._meta_to_lists()
+            lons, lats = all_meta["longitude"], all_meta["latitude"]
+
+        else:
+            stat_data = subset.to_station_data_all(var_name, start, stop, ts_type)
+
+            if len(stat_data["stats"]) == 0:
+                raise DataCoverageError(
+                    "No stations could be found for input specs (var, start, stop, freq)"
+                )
+            lons = stat_data["longitude"]
+            lats = stat_data["latitude"]
+        if "label" not in kwargs:
+            kwargs["label"] = info_str
+
+        ax = plot_coordinates(
+            lons,
+            lats,
+            color=color,
+            marker=marker,
+            markersize=markersize,
+            legend=legend,
+            **kwargs,
+        )
+
+        if "title" in kwargs:
+            title = kwargs["title"]
+        else:
+            title = info_str
+        if add_title:
+            ax.set_title(title, fontsize=fontsize_base + 4)
+        return ax
